* Cleaned up Makefile, one can now build for a specific platform by issuing command 'make TARGET=xxx'. * Entire game is now compiled with g++ exclusively. * New optional Makefile directive USE_NO_EXCEPTIONS and one small modification to CML library allows compilation with -fno-exceptions, decreasing unnecessary code bloat and possible speed hit. * New optional Makefile directive USE_FLTO allows compilation using gcc's link-time-optimization feature. * New optional Makefile directive USE_INDEXED_BATCH will tell batch-drawing code to use indexed triangle lists to batch draw-calls instead of using degenerate triangle strips (is default). * Added list of pre-configured targets to top of Makefile (need to get Pickle to tell me what rdu platform is). * Added new performance-oriented directives to platforms in Makefile that would benefit. * SDLPort/eglport.c renamed to SDLPort/eglport.cpp to allow compilation of entire source using g++. * unrarlib.c renamed to unrarlib.cpp to allow compilation of entire source using g++. * Fixes to type of size parameter to calls to urarlib_get() to fix compilation warnings. * In unrarlib.cpp, added casts and changed misc parameters to fix all g++ compiler warnings. * Only include unrarlib.h when USE_UNRARLIB is defined.

+++ trunk/Hurrican/src/DX8Sound.cpp
@@ -342,7 +342,8 @@ void SoundManagerClass::LoadSong(const string &filename, int nr, bool loop /*=tr
 
     bool         fromrar  = false;
     char         *pData   = NULL;
-    unsigned int buf_size = 0;
+    //DKS - Changed from unsigned int to unsigned long to fix unrarlib compilation error:
+    unsigned long buf_size = 0;
     string       fullpath;
 
     // Zuerst checken, ob sich der Song in einem MOD-Ordner befindet
@@ -638,7 +639,8 @@ void SoundManagerClass::LoadWave(const string &filename, int nr, bool looped)
 
     bool         fromrar = false;
     char         *pData = NULL;
-    unsigned int buf_size = 0;
+    //DKS - Changed from unsigned int to unsigned long to fix unrarlib compilation error:
+    unsigned long buf_size = 0;
     string       fullpath;
 
     if (sounds[nr].data)
